sstv/requirements.txt
cffi==1.17.1
numpy==2.1.3
pillow==11.0.0
pycparser==2.22
PySSTV==0.5.7
scipy==1.14.1
soundfile==0.12.1
pydub==0.25.1